backend/requirements.txt
# See GDAL_INSTALL.md for instructions on how to setup Gdal and geoserver-rest
affine==2.3.0
alabaster==0.7.12
amqp==2.6.1
asgiref==3.2.10
attrs==20.2.0
Babel==2.9.0
billiard==3.6.3.0
boto3==1.14.53
botocore==1.17.53
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==5.2.1
croniter==0.3.36
cycler==0.11.0
Django==3.1
django-colorful==1.3
django-cors-headers==3.5.0
django-filter==2.3.0
django-leaflet==0.27.1
#django-raster==0.8
django-rq==2.4.0
djangorestframework==3.11.1
djangorestframework-gis==0.15
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
docutils==0.15.2
earthengine-api==0.1.376
enum34==1.1.10
fcm-django==1.0.12
Fiona==1.8.16
firebase-admin==5.3.0
fonttools==4.28.3
future==0.18.2
geopandas==0.8.1
# See GDAL_INSTALL.md for instructions on how to setup Gdal and geoserver-rest in dev environment. 
geoserver-rest==2.1.4
#google-api-core==2.11.0
google-api-core==2.12.0
google-api-python-client==1.12.11
#google-auth==2.16.0
google-auth==2.23.3
google-auth-httplib2==0.0.4
google-cloud-core==2.3.3
google-cloud-firestore==2.9.0
google-cloud-storage==2.10.0
google-crc32c==1.0.0
google-resumable-media==2.4.0
googleapis-common-protos==1.58.0
gunicorn==20.0.4
httplib2==0.18.1
httplib2shim==0.0.3
idna==2.10
imagesize==1.2.0
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.3.2
kombu==4.6.11
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.5.0
munch==2.5.0
natsort==7.0.1
#numpy==1.19.1
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
pip-autoremove==0.9.1
protobuf==3.19.6
psycopg2-binary==2.8.5
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pyparsing==2.4.7
pyproj==2.6.1.post1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rasterio==1.1.5
rasterstats==0.15.0
redis==3.5.3
requests==2.24.0
rq==1.6.1
rq-scheduler==0.10.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.2
sentry-sdk
setproctitle==1.1.10
setuptools-scm==6.3.2
Shapely==1.7.1
simplejson==3.17.2
six==1.15.0
snowballstemmer==2.0.0
snuggs==1.4.7
Sphinx==3.3.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
threadpoolctl==2.1.0
tomli==1.2.2
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
zipp==3.1.0
psutil==5.9.8
tabulate==0.9.0
python-dotenv
schedule==1.2.2
memory_profiler==0.61.0
